refactor: implement modular architecture with improved type safety
- Extract tools into resource-based modules (user, task, stream)
- Reduce main.ts complexity by 96% (1162 → 47 lines)
- Eliminate all 'any' types with proper TypeScript typing
- Add parameter destructuring for cleaner function signatures
- Create shared utilities for common patterns
- Move API documentation to dedicated resources
- Maintain 100% API compatibility while improving maintainability
